pyproject.toml
[project]
name = "novaseat-v2"
version = "0.1.0"
description = "AI-powered churn prevention system for NovaSeat B2B SaaS"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.11.0",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "joblib>=1.5.3",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "shap>=0.51.0",
    "xgboost>=3.2.0",
]
